analytics/requirements.txt
kafka-python==2.0.2
pandas==2.2.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.27
